pyproject.toml
[project]
name = "synthetic-join-datagen"
version = "0.1.0"
description = "Generates synthetic relation and join execution plan datasets"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bokeh>=3.8.0.dev1",
    "dask[complete]>=2025.5.1",
    "numpy>=2.3.0rc1",
    "pandas-stubs>=2.2.3.250527",
]
